+++ src/open_clip/tokenizer.py
@@ -187,15 +187,16 @@ def tokenize(texts: Union[str, List[str]], context_length: int = 77) -> torch.Lo
 
 class HFTokenizer:
     "HuggingFace tokenizer wrapper"
-    def __init__(self, tokenizer_name:str):
+    def __init__(self, tokenizer_name:str, context_length:int=77):
         from transformers import AutoTokenizer
         self.tokenizer = AutoTokenizer.from_pretrained(tokenizer_name)
+        self.context_length = context_length
 
-    def __call__(self, texts:Union[str, List[str]], context_length:int=77) -> torch.Tensor:
+    def __call__(self, texts:Union[str, List[str]]) -> torch.Tensor:
         # same cleaning as for default tokenizer, except lowercasing
         # adding lower (for case-sensitive tokenizers) will make it more robust but less sensitive to nuance
         if isinstance(texts, str):
             texts = [texts]
         texts = [whitespace_clean(basic_clean(text)) for text in texts]
-        input_ids = self.tokenizer(texts, return_tensors='pt', max_length=context_length, padding='max_length', truncation=True).input_ids
+        input_ids = self.tokenizer(texts, return_tensors='pt', max_length=self.context_length, padding='max_length', truncation=True).input_ids
         return input_ids