+++ src/training/params.py
@@ -38,7 +38,7 @@ def parse_args(args):
     )
     parser.add_argument(
         "--dataset-type",
-        choices=["webdataset", "csv", "synthetic", "auto"],
+        choices=["webdataset", "csv", "synthetic", "auto", "hf"],
         default="auto",
         help="Which type of dataset to process."
     )
@@ -65,6 +65,18 @@ def parse_args(args):
         type=str,
         default="title",
         help="For csv-like datasets, the name of the key for the captions."
+    )
+    parser.add_argument(
+        "--hf-img-key",
+        type=str,
+        default=None,
+        help="For huggingface datasets, the name of the key for the image paths."
+    )
+    parser.add_argument(
+        "--hf-caption-key",
+        type=str,
+        default=None,
+        help="For huggingface datasets, the name of the key for the captions."
     )
     parser.add_argument(
         "--imagenet-val",