+++ tests/test_hf_model.py
@@ -1,7 +1,7 @@
 import pytest
 
 import torch
-from open_clip.hf_model import _POOLERS, HFTextEncoder
+from open_clip.hf_model import _POOLERS, HFEncoder
 from transformers import AutoConfig
 from transformers.modeling_outputs import BaseModelOutput
 # test poolers
@@ -16,12 +16,12 @@ def test_poolers():
         res = pooler(x, mask)
         assert res.shape == (bs, d), f"{name} returned wrong shape"
 
-# test HFTextEncoder
+# test HFEncoder
 @pytest.mark.parametrize("model_id", ["arampacha/roberta-tiny", "roberta-base", "xlm-roberta-base", "google/mt5-base"])
 def test_pretrained_text_encoder(model_id):
     bs, sl, d = 2, 10, 64
     cfg = AutoConfig.from_pretrained(model_id)
-    model = HFTextEncoder(model_id, d, proj='linear')
+    model = HFEncoder(model_id, d, proj='linear')
     x = torch.randint(0, cfg.vocab_size, (bs, sl))
     with torch.no_grad():
         emb = model(x)